+++ CHANGES.rst
@@ -26,11 +26,21 @@ describe future plans.
 
    Release expected by 2024-12-31.
 
+   Enhancements
+   ------------
+
+   - Add 'dynamic_import()' (support 'ad_creator()' from device file).
+
    Fixes
    -----
 
-   - PVPositionerSoftDone used an invalid subscription event type
+   - 'PVPositionerSoftDone' used an invalid subscription event type
      in unusual cases (with fake ophyd simulated devices).
+
+   Maintenance
+   -----------
+
+   - In 'ad_creator()', convert text class name to class object.
 
 1.7.1
 ******